Python2 is an unholy mistake and should never be used
Imagine still using python2 in 2022
Added code, a lot of it. importing signal caused an unknown error that was not reported nor explained by this horrendous wallaby software. It is 3:23 am. This marks the second day I am going to sleep around 4am. The hubris and sloth of man (procrastination) is a horrible thing.